fix: replaced criterions  [ref:#1156]
