+++ apps/gauzy/src/app/@shared/candidate/candidate-interview-mutation/candidate-interview-mutation.component.ts
@@ -341,7 +341,7 @@ export class CandidateInterviewMutationComponent
 	}
 	route() {
 		this.dialogRef.close();
-		this.router.navigate(['/pages/employees/candidates/criterions']);
+		this.router.navigate(['/pages/employees/candidates/interviews']);
 	}
 
 	ngOnDestroy() {
