--- apps/gauzy/src/app/pages/candidates/candidate-criterions/candidate-criterions.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-	selector: 'ga-candidate-criterions',
-	templateUrl: './candidate-criterions.component.html',
-	styleUrls: ['./candidate-criterions.component.scss']
-})
-export class CandidateCriterionsComponent {
-	constructor() {}
-}
