+++ apps/gauzy/src/app/pages/candidates/candidates.component.ts
@@ -121,11 +121,6 @@ export class CandidatesComponent extends TranslationBaseComponent
 				title: this.getTranslation('CONTEXT_MENU.CANDIDATE_STATISTIC'),
 				icon: 'bar-chart-outline',
 				link: 'pages/employees/candidates/statistic'
-			},
-			{
-				title: this.getTranslation('CONTEXT_MENU.CANDIDATE_CRITERIONS'),
-				icon: 'options-2-outline',
-				link: 'pages/employees/candidates/criterions'
 			}
 		];
 	}
