+++ apps/gauzy/src/app/pages/candidates/edit-candidate/edit-candidate-profile/edit-candidate-feedbacks/edit-candidate-feedbacks.component.scss
@@ -215,7 +215,6 @@
       justify-content: flex-start;
       align-items: stretch;
       flex-wrap: wrap;
-      width: 48%;
       padding: 1rem;
 
       &-criterion {
