+++ apps/gauzy/src/app/pages/candidates/manage-candidate-interviews/interview-calendar /interview-calendar.component.html
@@ -1,0 +1,39 @@
+<nb-card-header>
+	<div class="main-header">
+		<h5>
+			{{
+				'CANDIDATES_PAGE.EDIT_CANDIDATE.INTERVIEW.SCHEDULED_INTERVIEWS'
+					| translate
+			}}
+		</h5>
+		<div class="buttons">
+			<ga-employee-multi-select
+				class="multi-select"
+				[allEmployees]="employees"
+				[label]="false"
+				(selectedChange)="onEmployeeSelected($event)"
+			>
+			</ga-employee-multi-select>
+			<ga-candidate-multi-select
+				class="multi-select"
+				[allCandidates]="candidates"
+				(selectedChange)="onCandidateSelected($event)"
+			>
+			</ga-candidate-multi-select>
+			<button nbButton status="primary" (click)="add()">
+				<nb-icon icon="plus-outline"></nb-icon>
+				{{
+					'CANDIDATES_PAGE.EDIT_CANDIDATE.INTERVIEW.ADD_INTERVIEW'
+						| translate
+				}}
+			</button>
+		</div>
+	</div>
+</nb-card-header>
+<nb-card-body>
+	<full-calendar
+		*ngIf="calendarOptions"
+		#calendar
+		[options]="calendarOptions"
+	></full-calendar>
+</nb-card-body>
