+++ apps/gauzy/src/app/pages/candidates/manage-candidate-interviews/interview-calendar /interview-calendar.component.scss
@@ -19,8 +19,3 @@
   height: 24px;
   margin-bottom: 5px;
 }
-::ng-deep {
-  .nb-theme-default .tab-link:focus {
-    outline: none !important;
-  }
-}
