+++ apps/gauzy/src/app/pages/candidates/manage-candidate-interviews/interview-criterions/candidate-personal-qualities/candidate-personal-qualities.component.scss
