+++ apps/gauzy/src/app/pages/candidates/manage-candidate-interviews/interview-criterions/candidate-technologies/candidate-technologies.component.scss
