+++ apps/gauzy/src/app/pages/candidates/manage-candidate-interviews/interview-criterions/interview-criterions.component.scss
@@ -5,7 +5,6 @@
   align-items: flex-start;
 
   &-card {
-    padding: 1rem;
     width: 48%;
   }
 }
