+++ apps/gauzy/src/assets/i18n/en.json
@@ -789,6 +789,12 @@
 			"NEXT": "Next",
 			"PREVIOUS": "Previous"
 		},
+		"MANAGE_INTERVIEWS": {
+			"CALENDAR": "Calendar",
+			"INTERVIEWS": "Interviews",
+			"MANAGE_INTERVIEWS": "Manage Interviews",
+			"CRITERIONS": "Criterions"
+		},
 		"STATISTIC": {
 			"CANDIDATE_STATISTICS": "Candidate statistics",
 			"CANDIDATE_RATING": "Candidate rating",
